refactor(x/data): update data module to use orm (part 2)
