+++ x/data/server/hasher/hasher_test.go
@@ -1,0 +1,98 @@
+package hasher
+
+import (
+	"fmt"
+	"hash"
+	"hash/fnv"
+	"math"
+	"testing"
+
+	"github.com/cosmos/cosmos-sdk/store/mem"
+	"github.com/stretchr/testify/require"
+
+	"github.com/tendermint/tendermint/libs/rand"
+)
+
+func TestHasher(t *testing.T) {
+	// test default case with good params
+	hasher, err := NewHasher(nil)
+	require.NoError(t, err)
+	testHasher(t, hasher, 5)
+
+	// test suboptimal case to trigger varint fallback
+	hasher, err = NewHasherWithOptions(HashOptions{
+		MinLength: 1,
+		NewHash: func() hash.Hash {
+			return sixteenBitHash{
+				fnv.New32(),
+			}
+		},
+	})
+	require.NoError(t, err)
+	testHasher(t, hasher, 5)
+}
+
+type sixteenBitHash struct {
+	hash.Hash
+}
+
+func (h sixteenBitHash) Sum(b []byte) []byte {
+	bz := h.Hash.Sum(b)
+	// just return b + the first two bytes
+	return bz[:len(b)+2]
+}
+
+func testHasher(t *testing.T, h Hasher, k int) {
+	hasher := h.(hasher)
+	store := mem.NewStore()
+	n := int(math.Pow10(k))
+	data := make([][]byte, n)
+	values := map[string]bool{}
+	totalCollisions := 0
+	secondaryCollisions := 0
+
+	for i := 0; i < n; i++ {
+		var value []byte
+		var valueStr string
+		for {
+			m := rand.Int31n(256)
+			value = rand.Bytes(int(m))
+			valueStr = fmt.Sprintf("%x", value)
+			if !values[valueStr] {
+				break
+			}
+		}
+		data[i] = value
+		values[valueStr] = true
+
+		id := []byte(nil)
+
+		for c := 0; ; c++ {
+			id = hasher.CreateID(value, c)
+			v := store.Get(id)
+			if len(v) == 0 {
+				store.Set(id, value)
+				break
+			}
+			totalCollisions += 1
+			if c > 1 {
+				secondaryCollisions += 1
+			}
+		}
+
+		newId := []byte(nil)
+		store = mem.NewStore()
+
+		for c := 0; ; c++ {
+			newId = hasher.CreateID(value, c)
+			v := store.Get(newId)
+			if len(v) == 0 {
+				store.Set(newId, value)
+				break
+			}
+			require.Equal(t, id, newId)
+		}
+	}
+
+	t.Logf("total collisions: %d / %.0e, secondary collisions: %d, collision rate: %.4f%%", totalCollisions, float64(n), secondaryCollisions, float64(totalCollisions)/float64(n)*100.0)
+}
