+++ x/data/server/utils.go
@@ -4,25 +4,19 @@ import (
 	"context"
 
 	gogotypes "github.com/gogo/protobuf/types"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
 	"github.com/regen-network/regen-ledger/x/data"
 )
 
-func (s serverImpl) getID(ctx context.Context, iri string) ([]byte, error) {
-	store := sdk.UnwrapSDKContext(ctx).KVStore(s.storeKey)
-	id := s.iriIDTable.GetID(store, []byte(iri))
-	if len(id) == 0 {
-		return nil, status.Errorf(codes.NotFound, "can't find %s", iri)
+func (s serverImpl) getEntry(ctx context.Context, id []byte) (*data.ContentEntry, error) {
+	dataId, err := s.stateStore.DataIDTable().Get(ctx, id)
+	if err != nil {
+		return nil, err
 	}
 
-	return id, nil
-}
-
-func (s serverImpl) getEntry(ctx context.Context, store sdk.KVStore, id []byte) (*data.ContentEntry, error) {
+	contentHash, err := data.ParseIRI(dataId.Iri)
+	if err != nil {
+		return nil, err
+	}
 
 	dataAnchor, err := s.stateStore.DataAnchorTable().Get(ctx, id)
 	if err != nil {
@@ -34,16 +28,10 @@ func (s serverImpl) getEntry(ctx context.Context, store sdk.KVStore, id []byte) 
 		Nanos:   dataAnchor.Timestamp.Nanos,
 	}
 
-	iri := string(s.iriIDTable.GetValue(store, id))
-	contentHash, err := data.ParseIRI(iri)
-	if err != nil {
-		return nil, err
-	}
-
 	entry := &data.ContentEntry{
-		Timestamp: timestamp,
-		Iri:       iri,
 		Hash:      contentHash,
+		Iri:       dataId.Iri,
+		Timestamp: timestamp,
 	}
 
 	return entry, nil
